+++ src/gun.js
@@ -1,4 +1,4 @@
-import BuyItem from "./buy-item.js";
+import BuyItem from './buy-item.js';
 
 export default class Gun extends BuyItem {
