+++ src/modules/Gun/assets/guns/10g.png
