+++ src/modules/Gun/assets/guns/1g.png
