+++ src/modules/Gun/assets/guns/2g.png
