+++ src/modules/Gun/assets/guns/3g.png
