+++ src/modules/Gun/assets/guns/4g.png
