+++ src/modules/Gun/assets/guns/5g.png
