+++ src/modules/Gun/assets/guns/6g.png
