+++ src/modules/Gun/assets/guns/7g.png
