+++ src/modules/Gun/assets/guns/8g.png
