--- src/modules/Penguin/constants/index.js
@@ -1,1 +1,0 @@
-export { bodiesMap, gunsMap } from './assetsMaps.js';
