--- src/modules/Penguin/index.js
@@ -1,4 +1,0 @@
-export { Penguin } from './Penguin.js';
-export { SampleScene } from './SampleScene.js';
-export { bodiesMap, gunsMap } from './constants/index.js';
-export { loadAssets } from './utils/loadAssets.js';
