--- src/modules/Penguin/utils/index.js
@@ -1,2 +1,0 @@
-export { importsToMap } from "./importsToMap";
-export { loadAssets } from "./loadAssets";
