+++ src/modules/Penguin/utils/loadAssets.js
@@ -1,4 +1,6 @@
-import { bodiesMap, gunsMap } from "../constants";
+
+import { gunsMap } from 'src/modules/Gun/constants/assetMap.js';
+import { bodiesMap } from 'src/modules/Penguin/constants/assetMap.js';
 
 /**
  * @param {Phaser.Scene} scene
@@ -7,10 +9,10 @@ import { bodiesMap, gunsMap } from "../constants";
  * @param {Record<string, string>} [options.guns = gunsMap]
  */
 export const loadAssets = (
-  scene,
-  { bodies = bodiesMap, guns = gunsMap } = {}
-) => {
-  Object.entries({ ...bodies, ...guns }).forEach(([key, value]) => {
-    scene.load.image(key, value);
-  });
+    scene,
+    { bodies = bodiesMap, guns = gunsMap } = {}
+) => { 
+    Object.entries({ ...bodies, ...guns }).forEach(([key, value]) => {
+        scene.load.image(key, value);
+    });
 };
