--- src/modules/PenguinsShop/index.js
@@ -1,3 +1,0 @@
-export { ShopItemCard } from "./Entities/ShopItemCard.js";
-export { PenguinsShop } from "./PenguinsShop.js";
-export { SampleScene } from "./SampleScene.js";
